+++ lib/recipes/lifespan-development/recipe.rb
@@ -1,0 +1,44 @@
+# frozen_string_literal: true
+
+LIFESPAN_DEVELOPMENT_RECIPE = Kitchen::BookRecipe.new(book_short_name: :lifespan_development) \
+do |doc, _resources|
+  include Kitchen::Directions
+
+  book = doc.book
+  metadata = book.metadata
+
+  book.search('cnx-pi').trash
+
+  BakePreface.v1(book: book)
+
+  BakeUnnumberedFigure.v1(book: book)
+  BakeUnnumberedTables.v1(book: book)
+
+  BakeChapterTitle.v1(book: book)
+  BakeChapterIntroductions.v1(book: book)
+
+  book.chapters.each do |chapter|
+    BakeNonIntroductionPages.v1(chapter: chapter)
+  end
+
+  # Appendix
+  book.pages('$.appendix').each do |page|
+    appendix_letter = [*('A'..'Z')][page.count_in(:book) - 1]
+
+    page.figures(only: :figure_to_number?).each do |figure|
+      BakeFigure.v1(figure: figure, number: "#{appendix_letter}#{figure.count_in(:page)}")
+    end
+
+    page.tables('$:not(.unnumbered)').each do |table|
+      BakeNumberedTable.v1(table: table, number: "#{appendix_letter}#{table.count_in(:page)}")
+    end
+    BakeAppendix.v1(page: page, number: appendix_letter)
+  end
+
+  BakeFootnotes.v1(book: book)
+  BakeToc.v1(book: book)
+  BakeLinkPlaceholders.v1(book: book)
+  BakeFolio.v1(book: book)
+  BakeRexWrappers.v1(book: book)
+  BakeLinks.v1(book: book)
+end